Created a command to change the prefix for a specific server
Use mongoose to save the id from the server and the prefix from that server to know what prefix it should use in a specific server. If the server doens't  have a register already in the DB a new one will be created in the first message, or when the bot is added to a new guild. When the bot is removed from a guild it deletes the register from the db.